WIP: update modem/ofono to latest, make it work with --src
I'm using this to test that branch:
http://git.sysmocom.de/ofono/log/?h=lynxis/osmodevcon_upstream_voice

In recent versions, ofono depends on the ell library. When building with
the external ell (the one packaged in Alpine), ofono fails to link with:

/usr/lib/gcc/x86_64-alpine-linux-musl/8.3.0/../../../../x86_64-alpine-linux-musl/bin/ld: src/main.o: in function `main':
main.c:(.text.startup+0x10f): undefined reference to `__stop___ell_debug'
/usr/lib/gcc/x86_64-alpine-linux-musl/8.3.0/../../../../x86_64-alpine-linux-musl/bin/ld: main.c:(.text.startup+0x116): undefined reference to `__start___ell_debug'
collect2: error: ld returned 1 exit status

This also happens, when forking Alpine's ell package and building ell
master. So I have changed it for now, to use the ell source from the
ofono source dir.

Currently, this package can be built with:
* cloning ofono and checking out lynxis/osmodevcon_upstream_voice
* applying Alpine's fix-TEMP_FAILURE_RETRY.patch to the ofono repo
* cloning ell master into the ofono dir
* pmbootstrap build --src=/home/user/code/ofono ofono