cross/*-armhf: rebuild packages because of new hostspec
Commit 915c7d8 ("functions: adjust armhf triplet") in abuild changes the
hostspec for armhf from armv6-alpine-linux-muslgnueabihf to
armv6-alpine-linux-musleabihf and as our cross packages use the hostspec
directly, this resulted in a discrepancy between what abuild says and
what is in our package. Rebuild the armhf packages for that - and adjust
some of the APKBUILDs to not use the $pkgrel variable directly.